+++ cli/cli.go
@@ -31,6 +31,7 @@ import (
 	"github.com/NethermindEth/sedge/configs"
 	"github.com/NethermindEth/sedge/internal/pkg/clients"
 	"github.com/NethermindEth/sedge/internal/pkg/generate"
+	"github.com/NethermindEth/sedge/internal/ui"
 	"github.com/NethermindEth/sedge/internal/utils"
 	"github.com/spf13/cobra"
 )
@@ -47,6 +48,7 @@ var (
 	network           string
 	feeRecipient      string
 	jwtPath           string
+	graffiti          string
 	install           bool
 	run               bool
 	y                 bool
@@ -57,6 +59,7 @@ var (
 	vlExtraFlags      *[]string
 	mapAllPorts       bool
 	noMev             bool
+	noValidator       bool
 )
 
 const (
@@ -120,6 +123,12 @@ func preRunCliCmd(cmd *cobra.Command, args []string) error {
 	} else if !utils.ContainsOnly(*services, []string{execution, consensus, validator}) {
 		return fmt.Errorf(configs.RunClientsError, strings.Join(*services, ","), strings.Join([]string{execution, consensus, validator}, ","))
 	}
+	// Exclude validator from run-clients if no-validator flag is set
+	if noValidator && utils.Contains(*services, validator) {
+		*services = utils.Filter(*services, func(s string) bool {
+			return s != validator
+		})
+	}
 
 	// Validate network
 	networks, err := utils.SupportedNetworks()
@@ -150,11 +159,6 @@ func preRunCliCmd(cmd *cobra.Command, args []string) error {
 		validatorParts := strings.Split(validatorName, ":")
 		validatorName = validatorParts[0]
 		validatorImage = strings.Join(validatorParts[1:], ":")
-	}
-
-	// Checkpoint Url
-	if checkpointSyncUrl == "" {
-		checkpointSyncUrl, _ = configs.NetworksToCheckpointUrl[network]
 	}
 
 	return nil
@@ -227,30 +231,46 @@ func runCliCmd(cmd *cobra.Command, args []string) []error {
 		}
 	}
 
+	combinedClients.Execution.Image = executionImage
+	combinedClients.Consensus.Image = consensusImage
+	combinedClients.Validator.Image = validatorImage
+	combinedClients.Validator.Omited = noValidator
+
 	// Generate docker-compose scripts
 	gd := generate.GenerationData{
-		ExecutionClient:   combinedClients.Execution.Name,
-		ExecutionImage:    executionImage,
-		ConsensusClient:   combinedClients.Consensus.Name,
-		ConsensusImage:    consensusImage,
-		ValidatorClient:   combinedClients.Validator.Name,
-		ValidatorImage:    validatorImage,
+		ExecutionClient:   combinedClients.Execution,
+		ConsensusClient:   combinedClients.Consensus,
+		ValidatorClient:   combinedClients.Validator,
 		GenerationPath:    generationPath,
 		Network:           network,
 		CheckpointSyncUrl: checkpointSyncUrl,
 		FeeRecipient:      feeRecipient,
 		JWTSecretPath:     jwtPath,
+		Graffiti:          graffiti,
 		FallbackELUrls:    *fallbackEL,
 		ElExtraFlags:      *elExtraFlags,
 		ClExtraFlags:      *clExtraFlags,
 		VlExtraFlags:      *vlExtraFlags,
 		MapAllPorts:       mapAllPorts,
-		Mev:               !noMev,
+		Mev:               !noMev && !noValidator,
 	}
-	elPort, clPort, err := generate.GenerateScripts(gd)
+	results, err := generate.GenerateScripts(gd)
 	if err != nil {
 		return []error{err}
 	}
+
+	// Clean generated .env and docker compose files
+	err = generate.CleanGenerated(results)
+	if err != nil {
+		return []error{err}
+	}
+
+	// Print final files
+	log.Infof(configs.CreatedFile, results.EnvFilePath)
+	ui.PrintFileContent(cmd.OutOrStdout(), results.EnvFilePath)
+
+	log.Infof(configs.CreatedFile, results.DockerComposePath)
+	ui.PrintFileContent(cmd.OutOrStdout(), results.DockerComposePath)
 
 	// If --run-clients=none was set then exit and don't run anything
 	if len(*services) == 0 {
@@ -276,33 +296,34 @@ func runCliCmd(cmd *cobra.Command, args []string) []error {
 		}
 	}
 
-	log.Info(configs.ValidatorTips)
+	if !noValidator {
+		log.Info(configs.ValidatorTips)
 
-	// TODO: Add ability to not run a validator at all
-	// Run validator after execution and consensus clients are synced, unless the user intencionally wants to run the validator service in the previous step
-	if !utils.Contains(*services, validator) {
-		// Wait for clients to start
-		//log.Info(configs.WaitingForNodesToStart)
-		//time.Sleep(waitingTime)
-		// Track sync of execution and consensus clients
-		// TODO: Parameterize wait arg of trackSync
-		if err = trackSync(monitor, elPort, clPort, time.Minute*5); err != nil {
-			return []error{err}
-		}
-
-		// TODO: Prompt for waiting for keystore and validator registration to run the validator
-		if run {
-			if err = runAndShowContainers([]string{validator}); err != nil {
+		// Run validator after execution and consensus clients are synced, unless the user intencionally wants to run the validator service in the previous step
+		if !utils.Contains(*services, validator) {
+			// Wait for clients to start
+			//log.Info(configs.WaitingForNodesToStart)
+			//time.Sleep(waitingTime)
+			// Track sync of execution and consensus clients
+			// TODO: Parameterize wait arg of trackSync
+			if err = trackSync(monitor, results.ELPort, results.CLPort, time.Minute*5); err != nil {
 				return []error{err}
 			}
-		} else {
-			// Let the user decide to see the instructions for executing the validator and exit or let the tool execute it
-			if err = RunValidatorOrExit(); err != nil {
-				return []error{err}
+
+			// TODO: Prompt for waiting for keystore and validator registration to run the validator
+			if run {
+				if err = runAndShowContainers([]string{validator}); err != nil {
+					return []error{err}
+				}
+			} else {
+				// Let the user decide to see the instructions for executing the validator and exit or let the tool execute it
+				if err = RunValidatorOrExit(); err != nil {
+					return []error{err}
+				}
 			}
 		}
+		log.Info(configs.HappyStaking)
 	}
-	log.Info(configs.HappyStaking)
 
 	return nil
 }
@@ -323,13 +344,17 @@ func init() {
 
 	cliCmd.Flags().StringVar(&checkpointSyncUrl, "checkpoint-sync-url", "", "Initial state endpoint (trusted synced consensus endpoint) for the consensus client to sync from a finalized checkpoint. Provide faster sync process for the consensus client and protect it from long-range attacks affored by Weak Subjetivity")
 
-	cliCmd.Flags().StringVarP(&network, "network", "n", "mainnet", "Target network. e.g. mainnet, prater, kiln, etc.")
+	cliCmd.Flags().StringVarP(&network, "network", "n", "mainnet", "Target network. e.g. mainnet, goerli, sepolia, etc.")
 
-	cliCmd.Flags().StringVar(&feeRecipient, "fee-recipient", "", "Suggested fee recipient. Is 20-byte Ethereum address which the execution layer might choose to set as the coinbase and the recipient of other fees or rewards. There is no guarantee that an execution node will use the suggested fee recipient to collect fees, it may use any address it chooses. It is assumed that an honest execution node will use the suggested fee recipient, but users should note this trust assumption")
+	cliCmd.Flags().StringVar(&feeRecipient, "fee-recipient", "", "Suggested fee recipient. Is a 20-byte Ethereum address which the execution layer might choose to set as the coinbase and the recipient of other fees or rewards. There is no guarantee that an execution node will use the suggested fee recipient to collect fees, it may use any address it chooses. It is assumed that an honest execution node will use the suggested fee recipient, but users should note this trust assumption")
 
 	cliCmd.Flags().BoolVar(&noMev, "no-mev-boost", false, "Not use mev-boost if supported")
 
+	cliCmd.Flags().BoolVar(&noValidator, "no-validator", false, "Exclude the validator from the full node setup. Designed for execution and consensus nodes setup without a validator node. Exclude also the validator from other flags. If set, mev-boost will not be used.")
+
 	cliCmd.Flags().StringVar(&jwtPath, "jwt-secret-path", "", "Path to the JWT secret file")
+
+	cliCmd.Flags().StringVar(&graffiti, "graffiti", "", "Graffiti to be used by the validator")
 
 	cliCmd.Flags().BoolVarP(&install, "install", "i", false, "Install dependencies if not installed without asking")