+++ internal/utils/networks_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 )
 
-var networks = []string{"mainnet", "kiln", "ropsten", "prater", "sepolia"}
+var networks = []string{"mainnet", "kiln", "ropsten", "goerli", "sepolia", "denver", "chiado", "gnosis"}
 
 func TestSupportedNetworks(t *testing.T) {
 	names, err := SupportedNetworks()